+++ src/main/java/com/praise/push/adapter/in/web/KeywordController.java
@@ -1,9 +1,8 @@
 package com.praise.push.adapter.in.web;
 
 import com.praise.push.application.port.in.dto.KeywordResponseDto;
-import com.praise.push.common.dto.ResponseDto;
+import com.praise.push.common.model.ResponseDto;
 import com.praise.push.application.port.in.KeywordUseCase;
-import com.praise.push.domain.Keyword;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
