+++ src/main/java/com/praise/push/common/error/model/ErrorResponseDto.java
@@ -1,4 +1,4 @@
-package com.praise.push.common.error.dto;
+package com.praise.push.common.error.model;
 
 import com.praise.push.common.ErrorCode;
 import lombok.Builder;
