+++ src/test/java/com/praise/push/application/KeywordServiceTest.java
@@ -2,6 +2,7 @@ package com.praise.push.application;
 
 import com.praise.push.application.port.in.dto.KeywordResponseDto;
 import com.praise.push.application.port.out.LoadKeywordPort;
+import com.praise.push.application.service.KeywordService;
 import com.praise.push.domain.Keyword;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
